+++ src/common/numa_topo.h
